+++ guillotina_cms/interfaces/content.py
@@ -2,10 +2,12 @@ from guillotina.interfaces import IFolder
 from guillotina.interfaces import IItem
 from guillotina_cms.fields.richtext import RichTextField
 from guillotina.fields import CloudFileField
+from guillotina_cms.directives import fieldset_field
 
 
 class IDocument(IFolder):
 
+    fieldset_field('text', 'default')
     text = RichTextField(
         title='Text',
         required=False,
@@ -13,6 +15,7 @@ class IDocument(IFolder):
 
 class IImage(IItem):
 
+    fieldset_field('image', 'default')
     image = CloudFileField(
         title='Image',
         required=False,
@@ -20,6 +23,7 @@ class IImage(IItem):
 
 class IFile(IItem):
 
+    fieldset_field('file', 'default')
     file = CloudFileField(
         title='File',
         required=False,
@@ -28,6 +32,7 @@ class IFile(IItem):
 
 class INews(IItem):
 
+    fieldset_field('file', 'default')
     text = RichTextField(
         title='Text',
         required=False,