+++ guillotina_cms/json/__init__.py
@@ -1,1 +1,2 @@
 from . import serialize_content  # noqa
+from . import serialize_schema  # noqa